Use the default coordinates of SVG and Canvas elements
Use the default coordindates, which is 0,0 at the top left corner
of the SVG and Canvas element. Update the coordinates of sample
seeds to accordingly. (Previously 0,0 is at center of the
elements).

This allows us to do away with accounting for the offsets when
translating mouse position to the grid coordinate.